Cleaning & upgrade. - The menus were moved from the cities to the regions. - New menus were added in Teldrassil. - Fixed a bug in the menu Muldore.
